ark-template/requirements.txt
asn1crypto==0.24.0
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
Django==1.11.4
django-node==4.0.0
django-rest-framework==0.1.0
django-rest-knox==3.1.3
django-webpack-loader==0.5.0
djangorestframework==3.6.4
idna==2.6
optional-django==0.1.0
pycparser==2.18
pyOpenSSL==17.5.0
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22

